+++ src/server/api/Package.js
@@ -11,8 +11,8 @@ const WareHouses = orm.import('../database/models/WareHouses')
 
 router.post('/Package/Count', jwt_decode({
   secret: secretKey
-}), (request, response) => {
-  Packages.count().then(count => {
+}), async (request, response) => {
+  await Packages.count().then(count => {
     response.json({
       count: count
     })
@@ -21,14 +21,14 @@ router.post('/Package/Count', jwt_decode({
 
 router.post('/Package/Max', jwt_decode({
   secret: secretKey
-}), (request, response) => {
-  Packages.count().then(count => {
+}), async (request, response) => {
+  await Packages.count().then(async count => {
     if (count === 0) {
       response.json({
         max: 0
       })
     } else {
-      Packages.max('package_id').then(max => {
+      await Packages.max('package_id').then(max => {
         response.json({
           max: max
         })
@@ -37,66 +37,77 @@ router.post('/Package/Max', jwt_decode({
   })
 })
 
+const Receptionists = orm.import('../database/models/Receptionists')
 router.post('/Package/Add', jwt_decode({
   secret: secretKey
-}), (request, response) => {
-  const payload = request.body
-  payload.sender_city = payload.sender_city[payload.sender_city.length - 1]
-  payload.receiver_city = payload.receiver_city[payload.receiver_city.length - 1]
-  Packages.findOne({
+}),async (request, response) => {
+  await Receptionists.findOne({
     where: {
-      package_id: payload.package_id
+      receptionist_id: request.user.uuid
     }
   })
-    .then( project => {
-      if (!project) {
-        let inSequence = false
-        Packages.count().then(count => {
-          Packages.max('package_id').then(max => {
-            inSequence = ((count === 0 && payload.package_id === 1) || payload.package_id === max + 1)
-            if (inSequence) {
-              payload.status = '已揽件'
-              Packages.create(payload)
-                .then(() => {
-                  response.sendStatus(200)
-                })
-            } else {
-              response.sendStatus(403)
-            }
-          })
-        })
-      } else {
-        Packages.update(payload,{
+    .then(async auth => {
+      if (auth != null) {
+        const payload = request.body
+        payload.sender_city = payload.sender_city[payload.sender_city.length - 1]
+        payload.receiver_city = payload.receiver_city[payload.receiver_city.length - 1]
+        await Packages.findOne({
           where: {
             package_id: payload.package_id
           }
-        }).then( () => {
-          response.sendStatus(200)
-          }
-        )
-          .catch( () => {
-            response.sendStatus(406)
+        })
+          .then(async project => {
+            if (!project) {
+              let inSequence = false
+              await Packages.count().then(async count => {
+                await Packages.max('package_id').then(async max => {
+                  inSequence = ((count === 0 && payload.package_id === 1) || payload.package_id === max + 1)
+                  if (inSequence) {
+                    payload.status = '已揽件'
+                    await Packages.create(payload)
+                      .then(() => {
+                        response.sendStatus(200)
+                      })
+                  } else {
+                    response.sendStatus(403)
+                  }
+                })
+              })
+            } else {
+              await Packages.update(payload,{
+                where: {
+                  package_id: payload.package_id
+                }
+              }).then( () => {
+                  response.sendStatus(200)
+                }
+              )
+                .catch( () => {
+                  response.sendStatus(406)
+                })
+            }
           })
+      } else {
+        response.sendStatus(403)
       }
     })
-
 })
 
 const Trackings = orm.import('../database/models/Trackings')
-router.post('/Package/Tracking', (request, response) => {
+router.post('/Package/Tracking', async (request, response) => {
   const params = request.body
-  Packages.findAll({
+  await Packages.findAll({
     where: params,
     attributes: ['status']
   })
-    .then(statusProject => {
+    .then(async statusProject => {
       if (statusProject.length === 0) {
         response.json({
           status: 'NOT FOUND'
         })
       } else {
         const status = statusProject[0].status
-        Trackings.findAll({
+        await Trackings.findAll({
           where: params,
           attributes: ['action', 'warehouse_id', 'transport_id', 'date'],
           order: [['date', 'DESC']],
@@ -116,27 +127,39 @@ router.post('/Package/Tracking', (request, response) => {
     })
 })
 
-router.post('/Package/Checkpoint', (request, response) => {
-  const params = request.body
-  WareHouses.findOne({
+const Transports = orm.import('../database/models/Transports')
+router.post('/Package/Checkpoint', async (request, response) => {
+  await Transports.findOne({
     where: {
-      warehouse_name: params.warehouse_name
-    },
-    attributes: ['warehouse_id']
+      transport_id: request.user.uuid
+    }
   })
-    .then(async warehouse => {
-      const warehouse_id = warehouse.get('warehouse_id')
-      await Trackings.create({
-        action: params.action,
-        date: Date.now(),
-        package_id: params.package_id,
-        warehouse_id: warehouse_id,
-        transport_id: request.user.uuid
-      }).then(() => {
-        response.sendStatus(200)
-      }, () => {
+    .then(async auth => {
+      if (auth != null) {
+        const params = request.body
+        await WareHouses.findOne({
+          where: {
+            warehouse_name: params.warehouse_name
+          },
+          attributes: ['warehouse_id']
+        })
+          .then(async warehouse => {
+            const warehouse_id = warehouse.get('warehouse_id')
+            await Trackings.create({
+              action: params.action,
+              date: Date.now(),
+              package_id: params.package_id,
+              warehouse_id: warehouse_id,
+              transport_id: request.user.uuid
+            }).then(() => {
+              response.sendStatus(200)
+            }, () => {
+              response.sendStatus(403)
+            })
+          })
+      } else {
         response.sendStatus(403)
-      })
+      }
     })
 })
 
@@ -144,13 +167,9 @@ const getCascadedLocation = require('../database/utils').getCascadedLocation
 
 router.post('/Package/Query', jwt_decode({
   secret: secretKey
-}), (request, response) => {
+}), async (request, response) => {
   const payload = request.body
-  console.log(payload)
-  console.log(request.user.uuid)
-  // 之后JWT生成token的时候加上组，这里取出组以后再做一次查权限
-  // 过期用插件自带的就好，不要自己做了
-  Packages.findAll(payload)
+  await Packages.findAll(payload)
     .then(async project => {
       for (const p of project){
         console.log(getCascadedLocation())
@@ -164,19 +183,27 @@ router.post('/Package/Query', jwt_decode({
 
 router.post('/Package/Delete', jwt_decode({
   secret: secretKey
-}), (request, response) => {
-  const params = request.body
-  console.log(request.user.uuid)
-  // 之后JWT生成token的时候加上组，这里取出组以后再做一次查权限
-  // 过期用插件自带的就好，不要自己做了
-  Packages.destroy({
-    where: params
+}), async (request, response) => {
+  await Receptionists.findOne({
+    where: {
+      receptionist_id: request.user.uuid
+    }
   })
-    .then(() => {
-      response.sendStatus(200)
-    })
-    .catch(() => {
-      response.sendStatus(406)
+    .then(async auth => {
+      if (auth != null) {
+        const params = request.body
+        await Packages.destroy({
+          where: params
+        })
+          .then(() => {
+            response.sendStatus(200)
+          })
+          .catch(() => {
+            response.sendStatus(406)
+          })
+      } else {
+        response.sendStatus(403)
+      }
     })
 })
